requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
Babel==2.10.1
beautifulsoup4==4.10.0
boto3==1.22.0
botocore==1.25.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
colorama==0.4.4
cryptography==36.0.2
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-heroku==0.3.1
django-storages==1.12.3
djangorestframework==3.13.1
filelock==3.4.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
jmespath==1.0.0
outcome==1.1.0
Pillow==9.0.1
platformdirs==2.4.1
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
s3transfer==0.5.2
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
tkcalendar==1.6.1
toml==0.10.2
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0
wsproto==1.1.0